Add documentation for actor_get_contacts_avatars_since. Changed test fixtures to include last avatar update time.
Fixed a bug where we would skip certain contacts:
The bug happens when we specify both limit and since_time.
It's possible that none of the contacts returned by the initial
limit query would fit the since_time criterion.

But even with the fix, we still can't handle users with a large number
of contacts. We should come up with a long term fix for that.



Review URL: http://rietku.appspot.com/20001

git-svn-id: http://jaikuengine.googlecode.com/svn/trunk@84 6315f19a-1029-11de-a3bf-55d8bd1d7a4a